backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
pyjwt==2.6.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
hypothesis==6.88.0
openai>=1.0.0
